remove guice
