+++ src/main/java/hudson/plugins/release/pipeline/ReleaseStep.java
@@ -1,18 +1,17 @@
 package hudson.plugins.release.pipeline;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.logging.Level;
+import java.util.Set;
 import java.util.logging.Logger;
 
 import javax.annotation.Nonnull;
 
-import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
-import org.jenkinsci.plugins.workflow.steps.AbstractStepExecutionImpl;
-import org.jenkinsci.plugins.workflow.steps.AbstractStepImpl;
 import org.jenkinsci.plugins.workflow.steps.Step;
 import org.jenkinsci.plugins.workflow.steps.StepContext;
-import org.jenkinsci.plugins.workflow.steps.StepContextParameter;
+import org.jenkinsci.plugins.workflow.steps.StepDescriptor;
+import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.jenkinsci.plugins.workflow.util.StaplerReferer;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.DoNotUse;
@@ -22,26 +21,20 @@ import org.kohsuke.stapler.DataBoundSetter;
 import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
 
-import com.google.inject.Inject;
+import com.google.common.collect.ImmutableSet;
 
-import hudson.AbortException;
 import hudson.Extension;
-import hudson.console.ModelHyperlinkNote;
-import hudson.model.Action;
+import hudson.FilePath;
 import hudson.model.AutoCompletionCandidates;
 import hudson.model.BuildableItem;
 import hudson.model.BuildableItemWithBuildWrappers;
-import hudson.model.Cause;
-import hudson.model.CauseAction;
 import hudson.model.ItemGroup;
 import hudson.model.Job;
 import hudson.model.ParameterDefinition;
 import hudson.model.ParameterValue;
-import hudson.model.Queue;
 import hudson.model.Run;
 import hudson.model.TaskListener;
 import hudson.plugins.release.ReleaseWrapper;
-import hudson.plugins.release.SafeParametersAction;
 import jenkins.model.Jenkins;
 import jenkins.model.ParameterizedJobMixIn;
 import net.sf.json.JSONArray;
@@ -53,7 +46,7 @@ import net.sf.json.JSONObject;
  * @author Alexey Merezhin
  * @since 2.7
  */
-public class ReleaseStep extends AbstractStepImpl {
+public class ReleaseStep extends Step {
     private static final Logger LOGGER = Logger.getLogger(ReleaseStep.class.getName());
 
     private String job;
@@ -64,13 +57,6 @@ public class ReleaseStep extends AbstractStepImpl {
     public ReleaseStep(String job) {
         this.job = job;
         parameters = new ArrayList<>();
-    }
-
-    private Object readResolve() {
-        if (parameters == null) {
-            parameters = new ArrayList<>();
-        }
-        return this;
     }
 
     public String getJob() {
@@ -94,85 +80,13 @@ public class ReleaseStep extends AbstractStepImpl {
         }
     }
 
-    public static class Execution extends AbstractStepExecutionImpl {
-        @StepContextParameter private transient Run<?,?> invokingRun;
-        @StepContextParameter private transient TaskListener listener;
-
-        @Inject(optional=true) transient ReleaseStep step;
-
-        private List<ParameterValue> updateParametersWithDefaults(ParameterizedJobMixIn.ParameterizedJob project,
-                List<ParameterValue> parameters) throws AbortException {
-
-            if (project instanceof BuildableItemWithBuildWrappers) {
-                ReleaseWrapper wrapper = ((BuildableItemWithBuildWrappers) project).getBuildWrappersList()
-                                                                                   .get(ReleaseWrapper.class);
-                if (wrapper != null) {
-                    for (ParameterDefinition pd : wrapper.getParameterDefinitions()) {
-                        boolean parameterExists = false;
-                        for (ParameterValue pv : parameters) {
-                            if (pv.getName()
-                                  .equals(pd.getName())) {
-                                parameterExists = true;
-                                break;
-                            }
-                        }
-                        if (!parameterExists) {
-                            parameters.add(pd.getDefaultParameterValue());
-                        }
-                    }
-                } else {
-                    throw new AbortException("Job doesn't have release plugin configuration");
-                }
-            }
-            return parameters;
-        }
-
-        @Override
-        public boolean start() throws Exception {
-            if (step.getJob() == null) {
-                throw new AbortException("Job name is not defined.");
-            }
-
-            final ParameterizedJobMixIn.ParameterizedJob project =
-                    Jenkins.getActiveInstance().
-                            getItem(step.getJob(), invokingRun.getParent(),
-                                    ParameterizedJobMixIn.ParameterizedJob.class);
-            if (project == null) {
-                throw new AbortException("No parametrized job named " + step.getJob() + " found");
-            }
-            listener.getLogger().println("Releasing project: " + ModelHyperlinkNote.encodeTo(project));
-
-            StepContext context = getContext();
-            LOGGER.log(Level.FINER, "scheduling a release of {0} from {1}", new Object[] { project, context });
-            Action[] actions = new Action[]{
-                    new ReleaseTriggerAction(context),
-                    new SafeParametersAction(updateParametersWithDefaults(project, step.getParameters())),
-                    new ReleaseWrapper.ReleaseBuildBadgeAction(),
-                    new CauseAction(new Cause.UpstreamCause(invokingRun))
-            };
-
-            Queue.Item item = ParameterizedJobMixIn.scheduleBuild2((Job<?, ?>) project, 0, actions);
-
-            if (item == null || item.getFuture() == null) {
-                throw new AbortException("Failed to trigger build of " + project.getFullName());
-            }
-
-            return false;
-        }
-
-        @Override
-        public void stop(@Nonnull Throwable cause) throws Exception {
-            getContext().onFailure(cause);
-        }
-
-        private static final long serialVersionUID = 1L;
+    @Override
+    public StepExecution start(StepContext stepContext) throws Exception {
+        return new ReleaseStepExecution(stepContext, this);
     }
 
     @Extension
-    public static class DescriptorImpl extends AbstractStepDescriptorImpl {
-        public DescriptorImpl() {
-            super(Execution.class);
-        }
+    public static class DescriptorImpl extends StepDescriptor {
 
         @Override
         public String getFunctionName() {
@@ -184,7 +98,13 @@ public class ReleaseStep extends AbstractStepImpl {
             return "Trigger release for the job";
         }
 
-        @Override public Step newInstance(StaplerRequest req, JSONObject formData) throws FormException {
+        @Override
+        public Set<Class<?>> getRequiredContext() {
+            return ImmutableSet.of(Run.class, TaskListener.class);
+        }
+
+        @Override
+        public Step newInstance(StaplerRequest req, JSONObject formData) throws FormException {
             ReleaseStep step = (ReleaseStep) super.newInstance(req, formData);
             // Cf. ParametersDefinitionProperty._doBuild:
             Object parameter = formData.get("parameter");
@@ -231,7 +151,7 @@ public class ReleaseStep extends AbstractStepImpl {
             return AutoCompletionCandidates.ofJobNames(ParameterizedJobMixIn.ParameterizedJob.class, value, context);
         }
 
-        @Restricted(DoNotUse.class) // for use from config.jelly
+        @Restricted(DoNotUse.class) // for use from config.jelly / parameters.groovy
         public String getContext() {
             Job<?,?> job = StaplerReferer.findItemFromRequest(Job.class);
             return job != null ? job.getFullName() : null;
